+++ app/src/main/java/com/example/android_app_sdvg/domain/entity/task/Task.kt
@@ -20,7 +20,7 @@ class Task(
     val dateStart: Long,
     val dateEnd: Long,
     val timer: Long,
-    val capacity: Long,
+    val capacity: Int,
     val periodicity: Int,
     val priority: Priority,
     val category: Category
